+++ src/Bootstrap5Generator/index.ts
@@ -10,6 +10,7 @@ export type BoostrapThemeSCSS = string | {
 export interface BootstrapTheme {
 	variables?: BootstrapVariables,
 	scss?: BoostrapThemeSCSS,
+	colors?: BootstrapVariables,
 	[key: string]: any
 }
 
@@ -28,6 +29,20 @@ export interface Bootstrap5GeneratorConstructor {
 	cache?: Cache,
 	ignoreMinify?: boolean
 }
+
+const defaultColorKeysHM: { [k: string]: boolean | undefined } = {
+	primary: true,
+	secondary: true,
+	success: true,
+	info: true,
+	warning: true,
+	danger: true,
+	light: true,
+	dark: true
+
+
+}
+
 export default class Bootstrap5Generator {
 
 	#cache: Cache | undefined
@@ -39,8 +54,63 @@ export default class Bootstrap5Generator {
 
 	}
 
+	#getVariableFromThemeColors(theme: BootstrapTheme): string {
+		if (!theme?.colors) return ""
+		return Object.keys(theme.colors).reduce((a: string, key: string) => (
+			`${a}
+			 $${key}:${theme!.colors![key]};`
+		), "")
+	}
+
+
+	#getThemeColors(theme: BootstrapTheme): string {
+		if (!theme?.colors) return ""
+		const defaultColorKeys = Object.keys(defaultColorKeysHM)
+		const keys = [...defaultColorKeys, ...Object.keys(theme.colors).filter(el => !defaultColorKeysHM[el])]
+		const themeColors = keys.reduce((a: string, key: string) => (
+			`${a}
+			 "${key}":$${key},`
+		), "")
+
+
+		const textColors = keys.reduce((a: string, key: string) => (
+			`${a}
+			 "${key}":shade-color($${key}, 60%),`
+		), "")
+
+		const subtleColors = keys.reduce((a: string, key: string) => (
+			`${a}
+			 "${key}": tint-color($${key}, 80%),`
+		), "")
+
+
+		const borderSubtleColors = keys.reduce((a: string, key: string) => (
+			`${a}
+			 "${key}": tint-color($${key}, 60%),`
+		), "")
+
+		return `
+		$theme-colors: (
+			${themeColors}
+		  );
+
+
+		  $theme-colors-text: (
+			${textColors}
+		  );
+
+		  $theme-colors-bg-subtle: (
+			${subtleColors}
+		  );
+
+		  $theme-colors-border-subtle: (
+			${borderSubtleColors}
+		  );
+		`
+	}
 
 	generateCSS(theme: BootstrapTheme): string {
+
 		const variables = theme.variables || {}
 		let scssBefore: string = ""
 		let scssAfter: string = ""
@@ -53,15 +123,68 @@ export default class Bootstrap5Generator {
 		const variableSCSS = Object.keys(variables).reduce((acc, val) => (
 			`${acc}${val}:${variables[val]};`
 		), "")
+
+
 		const result = sass.compileString(`
-				${variableSCSS}
 				${scssBefore} 
-				@import "bootstrap";
+
+				@import "functions";
+
+				${variableSCSS}
+				${this.#getVariableFromThemeColors(theme)}
+				@import "variables";
+				@import "variables-dark";
+
+
+				${this.#getThemeColors(theme)}
+				@import "maps";
+				@import "mixins";
+				@import "root";
+
+	
+				@import "reboot";
+				@import "type";
+				@import "images";
+				@import "containers";
+				@import "grid";
+				@import "helpers";
+
+
+				
+				@import "tables";
+				@import "forms";
+				@import "buttons";
+				@import "transitions";
+				@import "dropdown";
+				@import "button-group";
+				@import "nav";
+				@import "navbar";
+				@import "card";
+				@import "accordion";
+				@import "breadcrumb";
+				@import "pagination";
+				@import "badge";
+				@import "alert";
+				@import "progress";
+				@import "list-group";
+				@import "close";
+				@import "toasts";
+				@import "modal";
+				@import "tooltip";
+				@import "popover";
+				@import "carousel";
+				@import "spinners";
+				@import "offcanvas";
+				@import "placeholders";
+				
+				@import "utilities";
+
 				${scssAfter}
 			`, {
 			style: this.#ignoreMinify ? "expanded" : "compressed",
 			loadPaths: [`${__dirname}/bootstrap/scss`]
 		})
+
 
 		return result.css