+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added color support to theme generation, adding a new color will generate all related classes included the subtle colors (btn-new-color, ...)
 
 ## [0.3.0] - 2024-03-25
 - Added support to custom SCSS